requirements.txt
boto==2.49.0
dj-database-url==0.5.0
Django==2.1.2
django-storages==1.7.1
gunicorn==19.9.0
Pillow==5.3.0
psycopg2-binary==2.7.5
pytz==2018.5
whitenoise==4.1
